mozio/requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
pip==22.1.2
pytz==2022.1
PyYAML==6.0
setuptools==62.4.0
Shapely==1.8.2
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
wheel==0.37.1
